+++ src/main/java/study/till/back/dto/CreateCommonResponse.java
@@ -1,11 +1,11 @@
-package study.till.back.dto.post;
+package study.till.back.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class CreatePostReponse {
+public class CreateCommonResponse {
     private Long id;
     private String status;
     private String message;
